+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Changes
 - Fortran implementation of the transformation cost objective function
 added;
 - `trmobjfunc` supports planar angles as variables for optimization;
-- Function `trmplotminatomdist` to plot minimal interatomic distances; 
+- Functions `trmplotminatomdist` and `pdbplotminatomdist` to plot minimal interatomic 
+distances; 
 - Function `trmchangeangles` for convenient angle change in a model;
 - Option to leave only alpha carbon atoms added to `pdbbackbone`;
 - Option to consider planar angles added to `trmdistantangleindices`;